% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_pacman.R
\name{animate_pacman}
\alias{animate_pacman}
\title{animate_pacman}
\usage{
animate_pacman(
  pacman,
  ghosts = list(blinky, pinky, inky, clyde),
  file = NULL,
  caption = "&copy; Micka&euml;l '<i style='color:#21908CFF;'>Coeos</i>' Canouil",
  font_family = ""
)
}
\arguments{
\item{pacman}{something}

\item{ghosts}{something}

\item{file}{something}

\item{caption}{something}

\item{font_family}{something}
}
\value{
something
}
\description{
animate_pacman
}
