% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ghost_coord.R
\name{compute_ghost_coord}
\alias{compute_ghost_coord}
\title{Compute Ghost Coordinates}
\usage{
compute_ghost_coord(data)
}
\arguments{
\item{data}{A \code{data.frame} with the coordinates of Ghost moves.}
}
\value{
Returns a \code{tibble} of cartesian coordinates.
}
\description{
Compute Ghost Coordinates
}
