% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ghost_status.R
\name{compute_ghost_status}
\alias{compute_ghost_status}
\title{Compute Ghost Status}
\usage{
compute_ghost_status(ghost, pacman_moves, bonus_points_eaten)
}
\arguments{
\item{ghost}{A \code{data.frame} with the coordinates of Ghost moves.}

\item{pacman_moves}{Pac-Man computed moves.}

\item{bonus_points_eaten}{Computed datga for points eaten by Pac-Man.}
}
\value{
Returns a \code{tibble} with ghost status as a column along side coordinates.
}
\description{
Compute a ghost status, \emph{i.e.}, normal, weak or eaten.
}
