% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pacman_coord.R
\name{compute_pacman_coord}
\alias{compute_pacman_coord}
\title{Compute Pac-Man Coordinates}
\usage{
compute_pacman_coord(data)
}
\arguments{
\item{data}{A \code{data.frame} with the coordinates of Pac-Man moves.}
}
\value{
Returns a \code{tibble} of cartesian coordinates and Pac-Man direction.
}
\description{
Compute Pac-Man Coordinates
}
