% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_points_eaten.R
\name{compute_points_eaten}
\alias{compute_points_eaten}
\title{Compute Point Eaten by Pac-Man}
\usage{
compute_points_eaten(bonus_points, pacman_moves)
}
\arguments{
\item{bonus_points}{Bonus points coordinates.}

\item{pacman_moves}{Pac-Man computed moves.}
}
\value{
Returns a \code{tibble} with the coordinates of point eaten by Pac-Man.
}
\description{
Compute Point Eaten by Pac-Man
}
