% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pacman-package.R
\docType{package}
\name{ggpacman-package}
\alias{ggpacman}
\alias{ggpacman-package}
\title{ggpacman: A 'ggplot2' and 'gganimate' Version of Pac-Man}
\description{
A funny coding challenge to reproduce the game Pac-Man using 'ggplot2' and 'gganimate'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mcanouil/pacman}
  \item Report bugs at \url{https://github.com/mcanouil/pacman/issues}
}

}
\author{
\strong{Maintainer}: Mickaël Canouil \email{***REMOVED***} (\href{https://orcid.org/0000-0002-3396-4549}{ORCID})

}
\keyword{internal}
